+++ lib/features/sales/domain/entities/item_master.dart
@@ -11,8 +11,8 @@ class ItemMasterEntity {
   final int serialNo;
   final String? tocatId;
   final String? touomId;
-  final int minStock;
-  final int maxStock;
+  final double minStock;
+  final double maxStock;
   final int includeTax;
   final String? remarks;
   final int statusActive;
@@ -24,12 +24,13 @@ class ItemMasterEntity {
   final int popItem;
   final String? bpom;
   final String? expDate;
-  final int? margin;
+  final double? margin;
   final int? memberDiscount;
   final int? multiplyOrder;
   final int mergeQuantity;
 
-  ItemMasterEntity({required this.docId,
+  ItemMasterEntity({
+    required this.docId,
     required this.createDate,
     required this.updateDate,
     required this.itemCode,
@@ -67,8 +68,8 @@ class ItemMasterEntity {
     int? serialNo,
     String? tocatId,
     String? touomId,
-    int? minStock,
-    int? maxStock,
+    double? minStock,
+    double? maxStock,
     int? includeTax,
     String? remarks,
     int? statusActive,
@@ -80,7 +81,7 @@ class ItemMasterEntity {
     int? popItem,
     String? bpom,
     String? expDate,
-    int? margin,
+    double? margin,
     int? memberDiscount,
     int? multiplyOrder,
     int? mergeQuantity,
@@ -150,36 +151,45 @@ class ItemMasterEntity {
     return ItemMasterEntity(
       docId: map['docId'] as String,
       createDate: DateTime.fromMillisecondsSinceEpoch(map['createDate'] as int),
-      updateDate: map['updateDate'] != null ? DateTime.fromMillisecondsSinceEpoch(map['updateDate'] as int) : null,
+      updateDate: map['updateDate'] != null
+          ? DateTime.fromMillisecondsSinceEpoch(map['updateDate'] as int)
+          : null,
       itemCode: map['itemCode'] as String,
       itemName: map['itemName'] as String,
       invItem: map['invItem'] as int,
       serialNo: map['serialNo'] as int,
       tocatId: map['tocatId'] != null ? map['tocatId'] as String : null,
       touomId: map['touomId'] != null ? map['touomId'] as String : null,
-      minStock: map['minStock'] as int,
-      maxStock: map['maxStock'] as int,
+      minStock: map['minStock'] as double,
+      maxStock: map['maxStock'] as double,
       includeTax: map['includeTax'] as int,
       remarks: map['remarks'] != null ? map['remarks'] as String : null,
       statusActive: map['statusActive'] as int,
       activated: map['activated'] as int,
       isBatch: map['isBatch'] as int,
-      internalCode_1: map['internalCode_1'] != null ? map['internalCode_1'] as String : null,
-      internalCode_2: map['internalCode_2'] != null ? map['internalCode_2'] as String : null,
+      internalCode_1: map['internalCode_1'] != null
+          ? map['internalCode_1'] as String
+          : null,
+      internalCode_2: map['internalCode_2'] != null
+          ? map['internalCode_2'] as String
+          : null,
       openPrice: map['openPrice'] as int,
       popItem: map['popItem'] as int,
       bpom: map['bpom'] != null ? map['bpom'] as String : null,
       expDate: map['expDate'] != null ? map['expDate'] as String : null,
-      margin: map['margin'] != null ? map['margin'] as int : null,
-      memberDiscount: map['memberDiscount'] != null ? map['memberDiscount'] as int : null,
-      multiplyOrder: map['multiplyOrder'] != null ? map['multiplyOrder'] as int : null,
+      margin: map['margin'] != null ? map['margin'] as double : null,
+      memberDiscount:
+          map['memberDiscount'] != null ? map['memberDiscount'] as int : null,
+      multiplyOrder:
+          map['multiplyOrder'] != null ? map['multiplyOrder'] as int : null,
       mergeQuantity: map['mergeQuantity'] as int,
     );
   }
 
   String toJson() => json.encode(toMap());
 
-  factory ItemMasterEntity.fromJson(String source) => ItemMasterEntity.fromMap(json.decode(source) as Map<String, dynamic>);
+  factory ItemMasterEntity.fromJson(String source) =>
+      ItemMasterEntity.fromMap(json.decode(source) as Map<String, dynamic>);
 
   @override
   String toString() {
@@ -189,63 +199,62 @@ class ItemMasterEntity {
   @override
   bool operator ==(covariant ItemMasterEntity other) {
     if (identical(this, other)) return true;
-  
-    return 
-      other.docId == docId &&
-      other.createDate == createDate &&
-      other.updateDate == updateDate &&
-      other.itemCode == itemCode &&
-      other.itemName == itemName &&
-      other.invItem == invItem &&
-      other.serialNo == serialNo &&
-      other.tocatId == tocatId &&
-      other.touomId == touomId &&
-      other.minStock == minStock &&
-      other.maxStock == maxStock &&
-      other.includeTax == includeTax &&
-      other.remarks == remarks &&
-      other.statusActive == statusActive &&
-      other.activated == activated &&
-      other.isBatch == isBatch &&
-      other.internalCode_1 == internalCode_1 &&
-      other.internalCode_2 == internalCode_2 &&
-      other.openPrice == openPrice &&
-      other.popItem == popItem &&
-      other.bpom == bpom &&
-      other.expDate == expDate &&
-      other.margin == margin &&
-      other.memberDiscount == memberDiscount &&
-      other.multiplyOrder == multiplyOrder &&
-      other.mergeQuantity == mergeQuantity;
+
+    return other.docId == docId &&
+        other.createDate == createDate &&
+        other.updateDate == updateDate &&
+        other.itemCode == itemCode &&
+        other.itemName == itemName &&
+        other.invItem == invItem &&
+        other.serialNo == serialNo &&
+        other.tocatId == tocatId &&
+        other.touomId == touomId &&
+        other.minStock == minStock &&
+        other.maxStock == maxStock &&
+        other.includeTax == includeTax &&
+        other.remarks == remarks &&
+        other.statusActive == statusActive &&
+        other.activated == activated &&
+        other.isBatch == isBatch &&
+        other.internalCode_1 == internalCode_1 &&
+        other.internalCode_2 == internalCode_2 &&
+        other.openPrice == openPrice &&
+        other.popItem == popItem &&
+        other.bpom == bpom &&
+        other.expDate == expDate &&
+        other.margin == margin &&
+        other.memberDiscount == memberDiscount &&
+        other.multiplyOrder == multiplyOrder &&
+        other.mergeQuantity == mergeQuantity;
   }
 
   @override
   int get hashCode {
     return docId.hashCode ^
-      createDate.hashCode ^
-      updateDate.hashCode ^
-      itemCode.hashCode ^
-      itemName.hashCode ^
-      invItem.hashCode ^
-      serialNo.hashCode ^
-      tocatId.hashCode ^
-      touomId.hashCode ^
-      minStock.hashCode ^
-      maxStock.hashCode ^
-      includeTax.hashCode ^
-      remarks.hashCode ^
-      statusActive.hashCode ^
-      activated.hashCode ^
-      isBatch.hashCode ^
-      internalCode_1.hashCode ^
-      internalCode_2.hashCode ^
-      openPrice.hashCode ^
-      popItem.hashCode ^
-      bpom.hashCode ^
-      expDate.hashCode ^
-      margin.hashCode ^
-      memberDiscount.hashCode ^
-      multiplyOrder.hashCode ^
-      mergeQuantity.hashCode;
+        createDate.hashCode ^
+        updateDate.hashCode ^
+        itemCode.hashCode ^
+        itemName.hashCode ^
+        invItem.hashCode ^
+        serialNo.hashCode ^
+        tocatId.hashCode ^
+        touomId.hashCode ^
+        minStock.hashCode ^
+        maxStock.hashCode ^
+        includeTax.hashCode ^
+        remarks.hashCode ^
+        statusActive.hashCode ^
+        activated.hashCode ^
+        isBatch.hashCode ^
+        internalCode_1.hashCode ^
+        internalCode_2.hashCode ^
+        openPrice.hashCode ^
+        popItem.hashCode ^
+        bpom.hashCode ^
+        expDate.hashCode ^
+        margin.hashCode ^
+        memberDiscount.hashCode ^
+        multiplyOrder.hashCode ^
+        mergeQuantity.hashCode;
   }
 }
